+++ source/Scrumbles.py
@@ -4,11 +4,11 @@ import sys
 
 
 
-import ctypes
-import sys
-
-if 'win' in sys.platform:
-         ctypes.windll.shcore.SetProcessDpiAwareness(1)
+# import ctypes
+# import sys
+#
+# if 'win' in sys.platform:
+#         ctypes.windll.shcore.SetProcessDpiAwareness(1)
 
 app = masterView.masterView()
